refactoring nodes
